% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{residuals.dsge_fit}
\alias{residuals.dsge_fit}
\title{Residuals from a fitted DSGE model}
\usage{
\method{residuals}{dsge_fit}(object, ...)
}
\arguments{
\item{object}{A \code{dsge_fit} object.}

\item{...}{Additional arguments (currently unused).}
}
\value{
A matrix of prediction errors.
}
\description{
Returns one-step-ahead prediction errors.
}
